+++ headers/VertexBuffer.hpp
@@ -9,11 +9,11 @@ class VertexBuffer
 private:
 	unsigned int m_ID;
 	bool m_IsBound;
-	unsigned int m_BufferSize = 0;
+	size_t m_BufferSize = 0;
 	
 public:
 	VertexBuffer() = default;
-	VertexBuffer(const void * p_Data, unsigned int p_Size, unsigned int p_Draw = GL_DYNAMIC_DRAW)
+	VertexBuffer(const void * p_Data, size_t p_Size, unsigned int p_Draw = GL_DYNAMIC_DRAW)
 	{
 		m_BufferSize = p_Size;
 		glGenBuffers(1, &m_ID);
@@ -21,7 +21,7 @@ public:
 		glBufferData(GL_ARRAY_BUFFER, p_Size, p_Data, p_Draw);
 		glBindBuffer(GL_ARRAY_BUFFER, 0);
 	}
-	unsigned int GetSize() const
+	size_t GetSize() const
 	{
 		return m_BufferSize;
 	}
